+++ prompt_styles/human-single-line.sh
@@ -49,7 +49,8 @@ fg_branch_status() {
 fg_branch_name() {
     local light_magenta="\\[\\e[95m\\]"
     local branch_name=$(git rev-parse --abbrev-ref HEAD 2> /dev/null)
-    local only_local_branch=$(git branch -a 2> /dev/null | egrep "remotes/origin/${branch_name}" | wc -l)
+    local remote_name=$(git rev-parse --abbrev-ref --symbolic-full-name @{u} 2> /dev/null | cut -d"/" -f1)
+    local only_local_branch=$(git branch -a 2> /dev/null | egrep "remotes/${remote_name}/${branch_name}" | wc -l)
     local light_green="\\[\\e[92m\\]"
     local none="\\[\\e[39m\\]"